requirements.txt
discord.py==2.3.2
aiohttp==3.9.4.post0
aiosqlite==0.19.0
Flask==2.2.5
